+++ tests/testthat/test-txtq.R
@@ -1,5 +1,3 @@
-context("txtq")
-
 test_that("core txtq API works", {
   q <- txtq(tempfile())
   expect_true(file.exists(q$path()))
@@ -81,7 +79,7 @@ test_that("reset()", {
 })
 
 test_that("clean()", {
-  df <- function(index){
+  df <- function(index) {
     data.frame(
       title = as.character(index),
       message = as.character(letters[index]),
@@ -96,7 +94,7 @@ test_that("clean()", {
   expect_equal(q$log()[, cols], df(index = 1:5))
   expect_equal(q$count(), 3)
   expect_equal(q$total(), 5)
-  for (i in 1:2){
+  for (i in 1:2) {
     q$clean()
     expect_equal(q$list()[, cols], df(index = 3:5))
     expect_equal(q$log()[, cols], df(index = 3:5))