+++ tests/testthat/test-validate.R
@@ -1,4 +1,8 @@
-context("validate")
+test_that("valid txtq validates fine", {
+  q <- txtq(tempfile())
+  q$push(1, 2)
+  expect_silent(q$validate())
+})
 
 test_that("cannot create dir because a file is already there", {
   path <- tempfile()